+++ include/bitwuzla/cpp/parser.h
@@ -166,6 +166,13 @@ class Parser
    */
   std::ostream& diagnostic_output_stream() const;
 
+  /**
+   * Get current parser statistics.
+   * @return A map of strings of statistics entries, maps statistic name
+   *         to value.
+   */
+  std::map<std::string, std::string> statistics() const;
+
  private:
   std::unique_ptr<bzla::parser::Parser> d_parser;
 };